+++ src/CrossValidation/Reports/MulticlassBreakdown.php
@@ -4,6 +4,7 @@ namespace Rubix\ML\CrossValidation\Reports;
 
 use Rubix\ML\Estimator;
 use Rubix\ML\EstimatorType;
+use Rubix\ML\CrossValidation\Reports\Results\Report;
 use InvalidArgumentException;
 
 use function count;
@@ -20,7 +21,7 @@ use const Rubix\ML\EPSILON;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class MulticlassBreakdown implements Report
+class MulticlassBreakdown implements ReportGenerator
 {
     /**
      * The estimator types that this report is compatible with.
@@ -41,9 +42,9 @@ class MulticlassBreakdown implements Report
      * @param (string|int)[] $predictions
      * @param (string|int)[] $labels
      * @throws \InvalidArgumentException
-     * @return mixed[]
+     * @return \Rubix\ML\CrossValidation\Reports\Results\Report
      */
-    public function generate(array $predictions, array $labels) : array
+    public function generate(array $predictions, array $labels) : Report
     {
         if (count($predictions) !== count($labels)) {
             throw new InvalidArgumentException('Number of predictions'
@@ -162,9 +163,9 @@ class MulticlassBreakdown implements Report
             'cardinality' => $n,
         ];
 
-        return [
+        return new Report([
             'overall' => $overall,
             'classes' => $table,
-        ];
+        ]);
     }
 }