+++ tests/CrossValidation/Reports/ErrorAnalysisTest.php
@@ -3,8 +3,9 @@
 namespace Rubix\ML\Tests\CrossValidation\Reports;
 
 use Rubix\ML\EstimatorType;
-use Rubix\ML\CrossValidation\Reports\Report;
 use Rubix\ML\CrossValidation\Reports\ErrorAnalysis;
+use Rubix\ML\CrossValidation\Reports\Results\Report;
+use Rubix\ML\CrossValidation\Reports\ReportGenerator;
 use PHPUnit\Framework\TestCase;
 use Generator;
 
@@ -33,7 +34,7 @@ class ErrorAnalysisTest extends TestCase
     public function build() : void
     {
         $this->assertInstanceOf(ErrorAnalysis::class, $this->report);
-        $this->assertInstanceOf(Report::class, $this->report);
+        $this->assertInstanceOf(ReportGenerator::class, $this->report);
     }
 
     /**
@@ -58,9 +59,10 @@ class ErrorAnalysisTest extends TestCase
      */
     public function generate(array $predictions, array $labels, array $expected) : void
     {
-        $result = $this->report->generate($predictions, $labels);
+        $results = $this->report->generate($predictions, $labels);
 
-        $this->assertEquals($expected, $result);
+        $this->assertInstanceOf(Report::class, $results);
+        $this->assertEquals($expected, $results->toArray());
     }
 
     /**