+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Unreleased
+    - CV Report Generators now return Report objects
     - Allow hyphens and apostrophes in Word Tokenizer
     - Dataset conversion methods now return an Encoding object
     - Encodings are now writeable to disk